+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.3 (unreleased)
 
+* moved repository to newsdev
 * fixed an error that occurred when no keys were set for a group
+* fixed incorrect etcd address default for unset command
 
 ## 0.1.2